move pkg/build/{prune,cmd} into pkg/oc/cli/builds
