+++ hack/import-restrictions.json
@@ -463,9 +463,7 @@
       "github.com/openshift/origin/pkg/build/apis/build/install",
       "github.com/openshift/origin/pkg/build/client",
       "github.com/openshift/origin/pkg/build/client/internalversion",
-      "github.com/openshift/origin/pkg/build/cmd",
       "github.com/openshift/origin/pkg/build/controller/build/apis/defaults",
-      "github.com/openshift/origin/pkg/build/prune",
       "github.com/openshift/origin/pkg/build/registry/build",
       "github.com/openshift/origin/pkg/build/registry/buildconfig",
       "github.com/openshift/origin/pkg/build/util",
