+++ pkg/oc/admin/prune/builds.go
@@ -16,7 +16,7 @@ import (
 	buildapi "github.com/openshift/origin/pkg/build/apis/build"
 	buildclient "github.com/openshift/origin/pkg/build/client"
 	buildinternalclient "github.com/openshift/origin/pkg/build/generated/internalclientset"
-	"github.com/openshift/origin/pkg/build/prune"
+	"github.com/openshift/origin/pkg/oc/cli/builds/prune"
 	"github.com/openshift/origin/pkg/oc/cli/util/clientcmd"
 )
