+++ pkg/oc/cli/builds/OWNERS
@@ -1,0 +1,12 @@
+reviewers:
+  - bparees
+  - smarterclayton
+  - mfojtik
+  - soltysh
+  - csrwng
+  - gabemontero
+approvers:
+  - bparees
+  - smarterclayton
+  - mfojtik
+  - soltysh
