+++ pkg/oc/cli/builds/doc.go
@@ -1,2 +1,2 @@
 // Package cmd provides command helpers for builds
-package cmd
+package builds
