+++ pkg/oc/cli/builds/prune/data_test.go
