+++ pkg/oc/cli/builds/reaper.go
@@ -1,4 +1,4 @@
-package cmd
+package builds
 
 import (
 	"sort"
