+++ pkg/oc/cli/builds/reaper_test.go
@@ -1,4 +1,4 @@
-package cmd
+package builds
 
 import (
 	"fmt"
